+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `formatSwissSocialInsuranceNumber` string utility function
 
+### Fixed
+
+- `isValidSwissSocialInsuranceNumber` is now named properly
+
 ## [2.1.0] - 2025-09-03
 
 ### Added